Validate ouName path parameter with authorized token's
