+++ agent-manager-service/controllers/agent_configuration_controller.go
@@ -61,6 +61,10 @@ func (c *agentConfigurationController) CreateAgentModelConfig(w http.ResponseWri
 	projectName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	createdBy := "system"
 
 	// Bind request body
@@ -140,6 +144,10 @@ func (c *agentConfigurationController) GetAgentModelConfig(w http.ResponseWriter
 	agentName := r.PathValue(utils.PathParamAgentName)
 	configID := r.PathValue(utils.PathParamConfigId)
 
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	configUUID, err := uuid.Parse(configID)
 	if err != nil {
 		log.Error("GetAgentModelConfig: invalid config ID", "configId", configID, "error", err)
@@ -174,6 +182,10 @@ func (c *agentConfigurationController) ListAgentModelConfigs(w http.ResponseWrit
 	limit := getIntQueryParam(r, "limit", 20)
 	offset := getIntQueryParam(r, "offset", 0)
 
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	// Validate and clamp pagination parameters
 	if limit < 1 {
 		limit = 20
@@ -206,6 +218,10 @@ func (c *agentConfigurationController) UpdateAgentModelConfig(w http.ResponseWri
 	projectName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 	configID := r.PathValue(utils.PathParamConfigId)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 
 	configUUID, err := uuid.Parse(configID)
 	if err != nil {
@@ -274,6 +290,10 @@ func (c *agentConfigurationController) DeleteAgentModelConfig(w http.ResponseWri
 	projectName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 	configID := r.PathValue(utils.PathParamConfigId)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 
 	configUUID, err := uuid.Parse(configID)
 	if err != nil {
