+++ agent-manager-service/controllers/agent_controller.go
@@ -17,7 +17,6 @@
 package controllers
 
 import (
-	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -26,7 +25,6 @@ import (
 	"time"
 
 	"github.com/wso2/agent-manager/agent-manager-service/config"
-	"github.com/wso2/agent-manager/agent-manager-service/middleware"
 	"github.com/wso2/agent-manager/agent-manager-service/middleware/logger"
 	"github.com/wso2/agent-manager/agent-manager-service/services"
 	"github.com/wso2/agent-manager/agent-manager-service/spec"
@@ -177,25 +175,14 @@ func handleCommonErrors(w http.ResponseWriter, err error, fallbackMsg string) {
 	}
 }
 
-// validateOrgFromPath validates that the org in the path matches the caller's token org.
-func validateOrgFromPath(w http.ResponseWriter, ctx context.Context, pathOrg string) bool {
-	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
-	if !ok {
-		utils.WriteErrorResponse(w, http.StatusForbidden, "missing org context")
-		return false
-	}
-	if pathOrg != resolvedOrg.OuHandle {
-		utils.WriteErrorResponse(w, http.StatusForbidden, "org mismatch with token identity")
-		return false
-	}
-	return true
-}
-
 func (c *agentController) GetAgent(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
@@ -219,6 +206,9 @@ func (c *agentController) ListAgents(w http.ResponseWriter, r *http.Request) {
 	log := logger.GetLogger(ctx)
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 
 	// Parse query parameters
@@ -270,6 +260,9 @@ func (c *agentController) CreateAgent(w http.ResponseWriter, r *http.Request) {
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 
 	// Parse and validate request body
@@ -317,6 +310,9 @@ func (c *agentController) UpdateAgentBasicInfo(w http.ResponseWriter, r *http.Re
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
@@ -349,6 +345,9 @@ func (c *agentController) UpdateAgentBuildParameters(w http.ResponseWriter, r *h
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
@@ -381,6 +380,9 @@ func (c *agentController) GetAgentResourceConfigs(w http.ResponseWriter, r *http
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 	environment := r.URL.Query().Get("environment")
@@ -406,6 +408,9 @@ func (c *agentController) UpdateAgentResourceConfigs(w http.ResponseWriter, r *h
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 	environment := r.URL.Query().Get("environment")
@@ -442,6 +447,9 @@ func (c *agentController) DeleteAgent(w http.ResponseWriter, r *http.Request) {
 	log := logger.GetLogger(ctx)
 
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
@@ -460,6 +468,9 @@ func (c *agentController) BuildAgent(w http.ResponseWriter, r *http.Request) {
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
@@ -483,6 +494,9 @@ func (c *agentController) GetBuildLogs(w http.ResponseWriter, r *http.Request) {
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 	buildName := r.PathValue(utils.PathParamBuildName)
@@ -503,6 +517,9 @@ func (c *agentController) GetAgentRuntimeLogs(w http.ResponseWriter, r *http.Req
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
@@ -536,6 +553,9 @@ func (c *agentController) GetAgentMetrics(w http.ResponseWriter, r *http.Request
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
@@ -568,6 +588,9 @@ func (c *agentController) DeployAgent(w http.ResponseWriter, r *http.Request) {
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
@@ -607,6 +630,9 @@ func (c *agentController) ListAgentBuilds(w http.ResponseWriter, r *http.Request
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
@@ -659,6 +685,9 @@ func (c *agentController) GenerateName(w http.ResponseWriter, r *http.Request) {
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	// Parse and validate request body
 	var payload spec.ResourceNameRequest
 	if err := json.NewDecoder(r.Body).Decode(&payload); err != nil {
@@ -695,6 +724,9 @@ func (c *agentController) GetBuild(w http.ResponseWriter, r *http.Request) {
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 	buildName := r.PathValue(utils.PathParamBuildName)
@@ -716,6 +748,9 @@ func (c *agentController) GetAgentDeployments(w http.ResponseWriter, r *http.Req
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
@@ -736,6 +771,9 @@ func (c *agentController) UpdateDeploymentState(w http.ResponseWriter, r *http.R
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
@@ -787,6 +825,9 @@ func (c *agentController) GetAgentEndpoints(w http.ResponseWriter, r *http.Reque
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 	environment := r.URL.Query().Get("environment")
@@ -813,6 +854,9 @@ func (c *agentController) GetAgentConfigurations(w http.ResponseWriter, r *http.
 
 	// Extract path parameters
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
@@ -892,6 +936,9 @@ func (c *agentController) PublishKind(w http.ResponseWriter, r *http.Request) {
 	log := logger.GetLogger(ctx)
 
 	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
