+++ agent-manager-service/controllers/agent_kind_controller.go
@@ -53,6 +53,9 @@ func (c *agentKindController) ListKinds(w http.ResponseWriter, r *http.Request) 
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
 
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	limitStr := r.URL.Query().Get("limit")
 	if limitStr == "" {
 		limitStr = strconv.Itoa(utils.DefaultLimit)
@@ -85,6 +88,10 @@ func (c *agentKindController) GetKind(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	kindName := r.PathValue(utils.PathParamKindName)
 
 	result, err := c.kindService.GetKind(ctx, orgName, kindName)
@@ -100,6 +107,10 @@ func (c *agentKindController) UpdateKind(w http.ResponseWriter, r *http.Request)
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	kindName := r.PathValue(utils.PathParamKindName)
 
 	var payload spec.UpdateAgentKindRequest
@@ -121,6 +132,10 @@ func (c *agentKindController) DeleteKind(w http.ResponseWriter, r *http.Request)
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	kindName := r.PathValue(utils.PathParamKindName)
 
 	if err := c.kindService.DeleteKind(ctx, orgName, kindName); err != nil {
@@ -135,6 +150,10 @@ func (c *agentKindController) AddVersion(w http.ResponseWriter, r *http.Request)
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	kindName := r.PathValue(utils.PathParamKindName)
 
 	var payload spec.AddAgentKindVersionRequest
@@ -160,6 +179,10 @@ func (c *agentKindController) ListVersions(w http.ResponseWriter, r *http.Reques
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	kindName := r.PathValue(utils.PathParamKindName)
 
 	result, err := c.kindService.ListVersions(ctx, orgName, kindName)
@@ -175,6 +198,10 @@ func (c *agentKindController) GetVersion(w http.ResponseWriter, r *http.Request)
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	kindName := r.PathValue(utils.PathParamKindName)
 	versionTag := r.PathValue(utils.PathParamVersionTag)
 
@@ -191,6 +218,10 @@ func (c *agentKindController) DeleteVersion(w http.ResponseWriter, r *http.Reque
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	kindName := r.PathValue(utils.PathParamKindName)
 	versionTag := r.PathValue(utils.PathParamVersionTag)
 
@@ -206,6 +237,10 @@ func (c *agentKindController) ListKindAgents(w http.ResponseWriter, r *http.Requ
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	kindName := r.PathValue(utils.PathParamKindName)
 
 	result, err := c.kindService.ListKindAgents(ctx, orgName, kindName)
