+++ agent-manager-service/controllers/agent_token_controller.go
@@ -21,7 +21,6 @@ import (
 	"errors"
 	"net/http"
 
-	"github.com/wso2/agent-manager/agent-manager-service/middleware"
 	"github.com/wso2/agent-manager/agent-manager-service/middleware/jwtassertion"
 	"github.com/wso2/agent-manager/agent-manager-service/middleware/logger"
 	"github.com/wso2/agent-manager/agent-manager-service/services"
@@ -58,17 +57,10 @@ func (c *agentTokenController) GenerateToken(w http.ResponseWriter, r *http.Requ
 	projName := r.PathValue(utils.PathParamProjName)
 	agentName := r.PathValue(utils.PathParamAgentName)
 
-	// Validate org matches the caller's token
-	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
-	if !ok {
-		utils.WriteErrorResponse(w, http.StatusForbidden, "missing org context")
+    if !validateOrgFromPath(w, ctx, orgName) {
 		return
 	}
-	if orgName != resolvedOrg.OuHandle {
-		utils.WriteErrorResponse(w, http.StatusForbidden, "org mismatch with token identity")
-		return
-	}
-
+	
 	log.Info(
 		"GenerateToken request received",
 		"orgName", orgName,
