+++ agent-manager-service/controllers/helpers.go
@@ -1,0 +1,39 @@
+// Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com).
+//
+// WSO2 LLC. licenses this file to you under the Apache License,
+// Version 2.0 (the "License"); you may not use this file except
+// in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package controllers
+
+import (
+	"context"
+	"net/http"
+
+	"github.com/wso2/agent-manager/agent-manager-service/middleware"
+	"github.com/wso2/agent-manager/agent-manager-service/utils"
+)
+
+// validateOrgFromPath validates that the org in the path matches the caller's token org.
+func validateOrgFromPath(w http.ResponseWriter, ctx context.Context, pathOrg string) bool {
+	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
+	if !ok {
+		utils.WriteErrorResponse(w, http.StatusForbidden, "missing org context")
+		return false
+	}
+	if pathOrg != resolvedOrg.OuHandle {
+		utils.WriteErrorResponse(w, http.StatusForbidden, "invalid organization access")
+		return false
+	}
+	return true
+}
