+++ agent-manager-service/controllers/identity_controller.go
@@ -85,6 +85,11 @@ func (c *identityController) ListUsers(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
 	if !ok {
 		utils.WriteErrorResponse(w, http.StatusForbidden, "missing org context")
@@ -121,6 +126,12 @@ func (c *identityController) ListUsers(w http.ResponseWriter, r *http.Request) {
 func (c *identityController) GetUser(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	userID := r.PathValue(utils.PathParamUserID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -149,6 +160,11 @@ func (c *identityController) GetUser(w http.ResponseWriter, r *http.Request) {
 func (c *identityController) CreateUser(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
 	if !ok {
@@ -208,6 +224,12 @@ func (c *identityController) CreateUser(w http.ResponseWriter, r *http.Request) 
 func (c *identityController) UpdateUser(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	userID := r.PathValue(utils.PathParamUserID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -253,6 +275,12 @@ func (c *identityController) UpdateUser(w http.ResponseWriter, r *http.Request) 
 func (c *identityController) DeleteUser(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	userID := r.PathValue(utils.PathParamUserID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -291,6 +319,12 @@ func (c *identityController) DeleteUser(w http.ResponseWriter, r *http.Request) 
 func (c *identityController) GetUserGroups(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	userID := r.PathValue(utils.PathParamUserID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -326,6 +360,12 @@ func (c *identityController) GetUserGroups(w http.ResponseWriter, r *http.Reques
 func (c *identityController) GetUserRoles(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	userID := r.PathValue(utils.PathParamUserID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -364,6 +404,11 @@ func (c *identityController) GetUserRoles(w http.ResponseWriter, r *http.Request
 func (c *identityController) InviteUser(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
 	if !ok {
@@ -404,6 +449,11 @@ func (c *identityController) ListGroups(w http.ResponseWriter, r *http.Request) 
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
 	if !ok {
 		utils.WriteErrorResponse(w, http.StatusForbidden, "missing org context")
@@ -439,6 +489,12 @@ func (c *identityController) ListGroups(w http.ResponseWriter, r *http.Request) 
 func (c *identityController) GetGroup(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	groupID := r.PathValue(utils.PathParamGroupID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -469,6 +525,11 @@ func (c *identityController) CreateGroup(w http.ResponseWriter, r *http.Request)
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
 	if !ok {
 		utils.WriteErrorResponse(w, http.StatusForbidden, "missing org context")
@@ -498,6 +559,12 @@ func (c *identityController) CreateGroup(w http.ResponseWriter, r *http.Request)
 func (c *identityController) UpdateGroup(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	groupID := r.PathValue(utils.PathParamGroupID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -543,6 +610,12 @@ func (c *identityController) UpdateGroup(w http.ResponseWriter, r *http.Request)
 func (c *identityController) DeleteGroup(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	groupID := r.PathValue(utils.PathParamGroupID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -581,6 +654,12 @@ func (c *identityController) DeleteGroup(w http.ResponseWriter, r *http.Request)
 func (c *identityController) AddGroupMembers(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	groupID := r.PathValue(utils.PathParamGroupID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -627,6 +706,12 @@ func (c *identityController) AddGroupMembers(w http.ResponseWriter, r *http.Requ
 func (c *identityController) RemoveGroupMembers(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	groupID := r.PathValue(utils.PathParamGroupID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -673,6 +758,12 @@ func (c *identityController) RemoveGroupMembers(w http.ResponseWriter, r *http.R
 func (c *identityController) GetGroupMembers(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	groupID := r.PathValue(utils.PathParamGroupID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -714,6 +805,12 @@ func (c *identityController) GetGroupMembers(w http.ResponseWriter, r *http.Requ
 func (c *identityController) GetGroupRoles(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	groupID := r.PathValue(utils.PathParamGroupID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -755,6 +852,11 @@ func (c *identityController) ListRoles(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
 	if !ok {
 		utils.WriteErrorResponse(w, http.StatusForbidden, "missing org context")
@@ -791,6 +893,12 @@ func (c *identityController) ListRoles(w http.ResponseWriter, r *http.Request) {
 func (c *identityController) GetRole(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	roleID := r.PathValue(utils.PathParamRoleID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -822,6 +930,11 @@ func (c *identityController) CreateRole(w http.ResponseWriter, r *http.Request) 
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
 	if !ok {
 		utils.WriteErrorResponse(w, http.StatusForbidden, "missing org context")
@@ -852,6 +965,12 @@ func (c *identityController) CreateRole(w http.ResponseWriter, r *http.Request) 
 func (c *identityController) UpdateRole(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	roleID := r.PathValue(utils.PathParamRoleID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -902,6 +1021,12 @@ func (c *identityController) UpdateRole(w http.ResponseWriter, r *http.Request) 
 func (c *identityController) DeleteRole(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	roleID := r.PathValue(utils.PathParamRoleID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -944,6 +1069,12 @@ func (c *identityController) DeleteRole(w http.ResponseWriter, r *http.Request) 
 func (c *identityController) GetRoleAssignments(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	roleID := r.PathValue(utils.PathParamRoleID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -979,6 +1110,12 @@ func (c *identityController) GetRoleAssignments(w http.ResponseWriter, r *http.R
 func (c *identityController) AddRolePermissions(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	roleID := r.PathValue(utils.PathParamRoleID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -1019,6 +1156,12 @@ func (c *identityController) AddRolePermissions(w http.ResponseWriter, r *http.R
 func (c *identityController) RemoveRolePermissions(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	roleID := r.PathValue(utils.PathParamRoleID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -1059,6 +1202,12 @@ func (c *identityController) RemoveRolePermissions(w http.ResponseWriter, r *htt
 func (c *identityController) AddRoleAssignees(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	roleID := r.PathValue(utils.PathParamRoleID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
@@ -1099,6 +1248,12 @@ func (c *identityController) AddRoleAssignees(w http.ResponseWriter, r *http.Req
 func (c *identityController) RemoveRoleAssignees(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
+
+	orgName := r.PathValue(utils.PathParamOrgName)
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	roleID := r.PathValue(utils.PathParamRoleID)
 
 	resolvedOrg, ok := middleware.GetResolvedOrg(ctx)
