+++ agent-manager-service/controllers/llm_controller.go
@@ -110,6 +110,10 @@ func (c *llmController) CreateLLMProviderTemplate(w http.ResponseWriter, r *http
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
 
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	var req spec.CreateLLMProviderTemplateRequest
 	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
 		log.Error("CreateLLMProviderTemplate: failed to decode request", "error", err)
@@ -148,6 +152,10 @@ func (c *llmController) ListLLMProviderTemplates(w http.ResponseWriter, r *http.
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 
 	// Parse pagination parameters
 	limit := getIntQueryParam(r, "limit", 20)
@@ -190,6 +198,10 @@ func (c *llmController) GetLLMProviderTemplate(w http.ResponseWriter, r *http.Re
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	templateID := r.PathValue(utils.PathParamTemplateId)
 
 	template, err := c.templateService.Get(orgName, templateID)
@@ -217,6 +229,10 @@ func (c *llmController) UpdateLLMProviderTemplate(w http.ResponseWriter, r *http
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	templateID := r.PathValue(utils.PathParamTemplateId)
 
 	var req spec.UpdateLLMProviderTemplateRequest
@@ -269,6 +285,10 @@ func (c *llmController) DeleteLLMProviderTemplate(w http.ResponseWriter, r *http
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	templateID := r.PathValue(utils.PathParamTemplateId)
 
 	if err := c.templateService.Delete(orgName, templateID); err != nil {
@@ -318,6 +338,10 @@ func (c *llmController) CreateLLMProvider(w http.ResponseWriter, r *http.Request
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 
 	log.Info("CreateLLMProvider: starting", "orgName", orgName)
 
@@ -384,6 +408,10 @@ func (c *llmController) ListLLMProviders(w http.ResponseWriter, r *http.Request)
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
 
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
+
 	log.Info("ListLLMProviders: starting", "orgName", orgName)
 
 	// Parse pagination parameters
@@ -431,6 +459,10 @@ func (c *llmController) GetLLMProvider(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	providerID := r.PathValue(utils.PathParamProviderId)
 
 	log.Info("GetLLMProvider: starting", "orgName", orgName, "providerID", providerID)
@@ -476,6 +508,10 @@ func (c *llmController) UpdateLLMProvider(w http.ResponseWriter, r *http.Request
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	providerID := r.PathValue(utils.PathParamProviderId)
 
 	log.Info("UpdateLLMProvider: starting", "orgName", orgName, "providerID", providerID)
@@ -662,6 +698,10 @@ func (c *llmController) DeleteLLMProvider(w http.ResponseWriter, r *http.Request
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	providerID := r.PathValue(utils.PathParamProviderId)
 
 	log.Info("DeleteLLMProvider: starting", "orgName", orgName, "providerID", providerID)
@@ -700,6 +740,10 @@ func (c *llmController) CreateLLMProxy(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projectName := r.PathValue(utils.PathParamProjName)
 
 	// Resolve project name to UUID
@@ -761,6 +805,10 @@ func (c *llmController) ListLLMProxies(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projectName := r.PathValue(utils.PathParamProjName)
 
 	// Resolve project name to UUID
@@ -821,6 +869,10 @@ func (c *llmController) ListLLMProxiesByProvider(w http.ResponseWriter, r *http.
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	providerID := r.PathValue(utils.PathParamProviderId)
 
 	// Parse pagination parameters
@@ -873,6 +925,10 @@ func (c *llmController) GetLLMProxy(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projectName := r.PathValue(utils.PathParamProjName)
 	proxyID := r.PathValue(utils.PathParamProxyId)
 
@@ -914,6 +970,10 @@ func (c *llmController) UpdateLLMProxy(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projectName := r.PathValue(utils.PathParamProjName)
 	proxyID := r.PathValue(utils.PathParamProxyId)
 
@@ -979,6 +1039,10 @@ func (c *llmController) DeleteLLMProxy(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	projectName := r.PathValue(utils.PathParamProjName)
 	proxyID := r.PathValue(utils.PathParamProxyId)
 
@@ -1019,6 +1083,10 @@ func (c *llmController) UpdateLLMProviderCatalogStatus(w http.ResponseWriter, r 
 	log := logger.GetLogger(ctx)
 
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	providerID := r.PathValue(utils.PathParamProviderId)
 
 	// Decode request body
