+++ agent-manager-service/controllers/llm_deployment_controller.go
@@ -54,6 +54,10 @@ func (c *llmDeploymentController) DeployLLMProvider(w http.ResponseWriter, r *ht
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	providerID := r.PathValue(utils.PathParamProviderId)
 
 	log.Info("DeployLLMProvider: starting", "orgName", orgName, "providerID", providerID)
@@ -148,6 +152,10 @@ func (c *llmDeploymentController) UndeployLLMProviderDeployment(w http.ResponseW
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	providerID := r.PathValue(utils.PathParamProviderId)
 
 	// Parse query parameters
@@ -218,6 +226,10 @@ func (c *llmDeploymentController) RestoreLLMProviderDeployment(w http.ResponseWr
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	providerID := r.PathValue(utils.PathParamProviderId)
 
 	// Parse query parameters
@@ -287,6 +299,10 @@ func (c *llmDeploymentController) DeleteLLMProviderDeployment(w http.ResponseWri
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	providerID := r.PathValue(utils.PathParamProviderId)
 	deploymentID := r.PathValue(utils.PathParamDeploymentId)
 
@@ -338,6 +354,10 @@ func (c *llmDeploymentController) GetLLMProviderDeployment(w http.ResponseWriter
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	providerID := r.PathValue(utils.PathParamProviderId)
 	deploymentID := r.PathValue(utils.PathParamDeploymentId)
 
@@ -385,6 +405,10 @@ func (c *llmDeploymentController) GetLLMProviderDeployments(w http.ResponseWrite
 	ctx := r.Context()
 	log := logger.GetLogger(ctx)
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	providerID := r.PathValue(utils.PathParamProviderId)
 
 	// Parse optional query parameters
