+++ agent-manager-service/controllers/llm_proxy_apikey_controller.go
@@ -54,6 +54,10 @@ func (c *llmProxyAPIKeyController) CreateAPIKey(w http.ResponseWriter, r *http.R
 	log := logger.GetLogger(ctx)
 
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	proxyID := r.PathValue("id")
 
 	log.Info("CreateLLMProxyAPIKey: starting", "orgName", orgName, "proxyID", proxyID)
@@ -117,6 +121,10 @@ func (c *llmProxyAPIKeyController) RevokeAPIKey(w http.ResponseWriter, r *http.R
 	log := logger.GetLogger(ctx)
 
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	proxyID := r.PathValue("id")
 	keyName := r.PathValue("keyName")
 
@@ -150,6 +158,10 @@ func (c *llmProxyAPIKeyController) RotateAPIKey(w http.ResponseWriter, r *http.R
 	log := logger.GetLogger(ctx)
 
 	orgName := r.PathValue(utils.PathParamOrgName)
+
+	if !validateOrgFromPath(w, ctx, orgName) {
+		return
+	}
 	proxyID := r.PathValue("id")
 	keyName := r.PathValue("keyName")
